requirements.txt
# Interface
customtkinter>=5.2.0

# Editor Web (será usado na Etapa 2)
pywebview>=4.4.0

# Processamento HTML (será usado na Etapa 2)
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Geração DOCX (Etapa 3)
python-docx>=1.1.0
docxtpl>=0.16.0

# Modo Lote (Etapa 3)
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1

# Utilidades
Pillow>=10.0.0

# Testes
pytest>=7.4.0
pytest-cov>=4.1.0
